app/requirements.txt
paddlepaddle==2.4.2
paddleocr==2.6.1.3
opencv-python==4.6.0.66
numpy==1.21.6
scikit-image==0.19.3
streamlit
pillow
python-dotenv
